Rename crop.py to avoid name clash with the skimage.utils.crop function
add bounding_box_crop to skimage/utils/__init__.py
