+++ skimage/_shared/utils.py
@@ -5,6 +5,7 @@ import warnings
 
 import numpy as np
 
+from ..util.dtype import img_as_float
 from ._warnings import all_warnings, warn
