+++ skimage/util/__init__.py
@@ -1,21 +1,22 @@
 import functools
 import warnings
-import numpy as np
-from .dtype import (img_as_float32, img_as_float64, img_as_float,
-                    img_as_int, img_as_uint, img_as_ubyte,
-                    img_as_bool, dtype_limits)
-from .shape import view_as_blocks, view_as_windows
-from .noise import random_noise
-from .apply_parallel import apply_parallel
 
-from .arraycrop import crop
-from .compare import compare_images
-from ._regular_grid import regular_grid, regular_seeds
-from .unique import unique_rows
+import numpy as np
+
+from ._bounding_box_crop import bounding_box_crop
 from ._invert import invert
+from ._label import label_points
 from ._montage import montage
 from ._map_array import map_array
-from ._label import label_points
+from ._regular_grid import regular_grid, regular_seeds
+from .apply_parallel import apply_parallel
+from .arraycrop import crop
+from .compare import compare_images
+from .dtype import (dtype_limits, img_as_float, img_as_float32, img_as_float64,
+                    img_as_bool, img_as_int, img_as_ubyte, img_as_uint)
+from .noise import random_noise
+from .shape import view_as_blocks, view_as_windows
+from .unique import unique_rows
 
 
 __all__ = ['img_as_float32',
@@ -28,6 +29,7 @@ __all__ = ['img_as_float32',
            'dtype_limits',
            'view_as_blocks',
            'view_as_windows',
+           'bounding_box_crop',
            'crop',
            'compare_images',
            'map_array',
