+++ skimage/util/_bounding_box_crop.py
